add converter and fix smth

+++ src/main/java/com/melloProj/Mello/controllers/InProjectController.java
@@ -3,9 +3,11 @@ package com.melloProj.Mello.controllers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.melloProj.Mello.models.project.List;
 import com.melloProj.Mello.models.project.Task;
+import com.melloProj.Mello.models.user.MelloUser;
 import com.melloProj.Mello.services.project.ListService;
 import com.melloProj.Mello.services.project.ProjectService;
 import com.melloProj.Mello.services.project.TaskService;
+import com.melloProj.Mello.services.user.TokenService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,12 +25,20 @@ public class InProjectController {
     @Autowired
     ListService listService;
 
+    @Autowired
+    TokenService tokenService;
+
     @SneakyThrows
     @CrossOrigin
     @PostMapping("task")
     @Operation(summary = "Создать задачу")
     public ResponseEntity<String> postTask(@RequestParam("TOKEN") String token,
                                            @RequestBody Task task){
+        MelloUser user = tokenService.getUserByToken(token);
+        if(user == null){
+            return ResponseEntity.badRequest().body("Error: User is not found");
+        }
+
         return ResponseEntity.ok().body(new ObjectMapper().writeValueAsString(taskService.createTask(task)));
     }
     @SneakyThrows
@@ -37,15 +47,25 @@ public class InProjectController {
     @Operation(summary = "Удалить задачу")
     public ResponseEntity<String> deleteTask(@RequestParam("TOKEN") String token,
                                              @RequestBody Task task){
+        MelloUser user = tokenService.getUserByToken(token);
+        if(user == null){
+            return ResponseEntity.badRequest().body("Error: User is not found");
+        }
+
         return ResponseEntity.ok().body(new ObjectMapper().writeValueAsString(taskService.deleteTask(task.getId())));
     }
     @SneakyThrows
     @PostMapping("/{listId}/tasks/{taskId}")
     @Operation(summary = "Переместить задачу в новый список")
-    public ResponseEntity<Void> moveTask(@PathVariable List listId,
-                                         @PathVariable Long taskId,
-                                         @RequestBody List newList) {
-        listService.updateTaskList(taskId, newList);
+    public ResponseEntity<String> moveTask(@RequestParam("TOKEN") String token,
+                                           @PathVariable Long listId,
+                                           @PathVariable Long taskId) {
+        MelloUser user = tokenService.getUserByToken(token);
+        if(user == null){
+            return ResponseEntity.badRequest().body("Error: User is not found");
+        }
+
+        listService.updateTaskList(taskId, listId);
         return ResponseEntity.ok().build();
     }
 
@@ -55,6 +75,11 @@ public class InProjectController {
     @Operation(summary = "Получить все задачи")
     public ResponseEntity<String> getTasks(@RequestParam("TOKEN") String token,
                                           @PathVariable Long listId){
+        MelloUser user = tokenService.getUserByToken(token);
+        if(user == null){
+            return ResponseEntity.badRequest().body("Error: User is not found");
+        }
+
         return ResponseEntity.ok().body(new ObjectMapper().writeValueAsString(listService.getTasksByList(listService.getById(listId))));
     }
 
@@ -64,6 +89,11 @@ public class InProjectController {
     @Operation(summary = "Получить список задач")
     public ResponseEntity<String> getList(@RequestParam("TOKEN") String token,
                                           @PathVariable Long listId){
+        MelloUser user = tokenService.getUserByToken(token);
+        if(user == null){
+            return ResponseEntity.badRequest().body("Error: User is not found");
+        }
+
         return ResponseEntity.ok().body(new ObjectMapper().writeValueAsString(listService.getById(listId)));
     }
     @SneakyThrows
@@ -72,6 +102,11 @@ public class InProjectController {
     @Operation(summary = "Создать список задач")
     public ResponseEntity<String> postList(@RequestParam("TOKEN") String token,
                                            @RequestBody List list){
+        MelloUser user = tokenService.getUserByToken(token);
+        if(user == null){
+            return ResponseEntity.badRequest().body("Error: User is not found");
+        }
+
         return ResponseEntity.ok().body(new ObjectMapper().writeValueAsString(listService.createList(list)));
     }
     @SneakyThrows
@@ -79,7 +114,12 @@ public class InProjectController {
     @DeleteMapping("list")
     @Operation(summary = "Удалить список задач")
     public ResponseEntity<String> deleteList(@RequestParam("TOKEN") String token,
-                                             @RequestBody Task task){
-        return ResponseEntity.ok().body(new ObjectMapper().writeValueAsString(listService.deleteList(task.getId())));
+                                             @RequestBody Long taskId){
+        MelloUser user = tokenService.getUserByToken(token);
+        if(user == null){
+            return ResponseEntity.badRequest().body("Error: User is not found");
+        }
+
+        return ResponseEntity.ok().body(new ObjectMapper().writeValueAsString(listService.deleteList(taskId)));
     }
 }
